introduce and use a generic ContentDelta
